requirements.txt
requests==2.28.2
websocket-client==1.5.1
python-dateutil==2.8.2
